satcoin: control range with defines
When generating multiple CNFs from the current source, defines might come
in handy. Hence, add a define to create satisfiable formulas (-DSATCNF=X)
and unsatisfiable formulas (-DUNSATCNF=X), where X represents the size of
the range that is allowed to search for the actual nonce.

The range of valid nonces for the satisfiable formulas is always set up so
that the valid nonce is in the middle of the range. For unsatisfiable
formulas, the range starts exactly above the valid nonce.

The approach can be used by first compiling with goto-cc, and next creating
the formula with cbmc. The compilation with a default compiler is not
altered.